Move redirectLoggedIn to own module to avoid inflating login entrypoint
