+++ client/controller/web-util.js
@@ -2,23 +2,6 @@
  * External dependencies
  */
 import ReactDom from 'react-dom';
-import page from 'page';
-
-/**
- * Internal dependencies
- */
-import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
-
-export function redirectLoggedIn( context, next ) {
-	const userLoggedIn = isUserLoggedIn( context.store.getState() );
-
-	if ( userLoggedIn ) {
-		page( '/' );
-		return;
-	}
-
-	next();
-}
 
 export function render( context ) {
 	ReactDom.render( context.layout, document.getElementById( 'wpcom' ) );
