+++ client/login/redirect-logged-in/index.node.js
@@ -1,0 +1,8 @@
+export default function redirectLoggedIn( { isLoggedIn, res }, next ) {
+	if ( isLoggedIn ) {
+		res.redirect( '/' );
+		return;
+	}
+
+	next();
+}
