+++ client/login/redirect-logged-in/index.web.js
@@ -1,0 +1,20 @@
+/**
+ * External dependencies
+ */
+import page from 'page';
+
+/**
+ * Internal dependencies
+ */
+import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
+
+export default function redirectLoggedIn( context, next ) {
+	const userLoggedIn = isUserLoggedIn( context.store.getState() );
+
+	if ( userLoggedIn ) {
+		page( '/' );
+		return;
+	}
+
+	next();
+}
